Refactor: src/types 디렉토리 src/lib 내부로 옮김
src/types=>src/lib/types
