+++ src/lib/types/letterTypes.ts
